0.9.14 - Práce vue uploadery. Oprava zadavania dátumu pri platnosti článku.

+++ app/ApiModule/Presenters/DokumentyPresenter.php
@@ -10,7 +10,7 @@ use Nette\Utils\Strings;
 
 /**
  * Prezenter pre pristup k api dokumentov.
- * Posledna zmena(last change): 03.02.2022
+ * Posledna zmena(last change): 20.02.2022
  *
  * Modul: API
  *
@@ -18,7 +18,7 @@ use Nette\Utils\Strings;
  * @copyright  Copyright (c) 2012 - 2022 Ing. Peter VOJTECH ml.
  * @license
  * @link       http://petak23.echo-msz.eu
- * @version 1.0.0
+ * @version 1.0.1
  */
 class DokumentyPresenter extends BasePresenter {
 
@@ -26,6 +26,7 @@ class DokumentyPresenter extends BasePresenter {
   /** @var DbTable\Dokumenty @inject */
   public $documents;
 
+  /** @var String */
   public $wwwDir;
 
   public function __construct(array $parameters, String $wwwDir) {
@@ -44,27 +45,33 @@ class DokumentyPresenter extends BasePresenter {
   /** Uloženie dokumentu do DB 
    * @param int $id Id_hlavne_menu, ku ktorému ukladám dokument */
   public function actionSave(int $id) {
+    //$_post = json_decode(file_get_contents("php://input"), true);
     /* from POST:
     * - description
     * - type
     * - name
+    */
+    $values = $this->getHttpRequest()->getPost();
+
+    /* from POST:
     * - priloha
     * - thumb
     */
+    $files = $this->getHttpRequest()->getFiles();
     $user = $this->getUser();
     $hl_menu = $this->hlavne_menu->find($id);
-    try {
+    //try {
       $uloz = [ 
         'id_hlavne_menu'	 	=> $id,
         'id_user_main'      => $user->id,
         'id_user_roles'     => $hl_menu->id_user_roles,
-        'description'				=> isset($values->description) && strlen($values->description)>2 ? $values->description : NULL,
+        'description'				=> isset($values['description']) && strlen($values['description'])>2 ? $values['description'] : NULL,
         'change'						=> date("Y-m-d H:i:s", Time()),
-        'type'              => $values->type
+        'type'              => $values['type']
       ];
-      $name = isset($values->name) ? $values->name : "";
-      if ($values->priloha && $values->priloha->error == 0) { //Ak nahravam prilohu
-        $priloha_info = $this->_uploadPriloha($values);
+      $name = isset($values['name']) ? $values['name'] : "";
+      if ($files['priloha'] && $files['priloha']->error == 0) { //Ak nahravam prilohu
+        $priloha_info = $this->_uploadPriloha($values['id'], $files);
         $uloz = array_merge($uloz, [
           'name'				=> strlen($name)>2 ? $name : $priloha_info['finalFileName'],
           'web_name'  	=> Strings::webalize($priloha_info['finalFileName']),
@@ -73,46 +80,49 @@ class DokumentyPresenter extends BasePresenter {
           'thumb_file'	=> $priloha_info['thumb'],
           'type'        => $priloha_info['is_image'] ? 2 : $values->type
           ]);
-      } elseif ($values->thumb->hasFile() && $values->thumb->isImage()) { //Ak nahravam len nahlad
-        $uloz = array_merge($uloz, ['thumb_file'	=> $this->_uploadThumb($values)]);
+      } elseif ($files['thumb']->hasFile() && $files['thumb']->isImage()) { //Ak nahravam len nahlad
+        $uloz = array_merge($uloz, ['thumb_file'	=> $this->_uploadThumb($values['id'], $files['thumb'])]);
       }  else { //Ak len menim
         $uloz = array_merge($uloz, ['name' => strlen($name)>2 ? $name : ""]);
       }
-      $vysledok = $this->dokumenty->uloz($uloz, $values->id);
+      $vysledok = $this->documents->uloz($uloz, $values['id']);
       if (!empty($vysledok) && isset($priloha_info['is_image']) && $priloha_info['is_image']) { 
-        $this->dokumenty->oprav($vysledok['id'], ['znacka'=>'#I-'.$vysledok['id'].'#']);
+        $this->documents->oprav($vysledok['id'], ['znacka'=>'#I-'.$vysledok['id'].'#']);
       }
-		} catch (Database\DriverException $e) {
+      if ($this->isAjax()) {
+        //
+        $this->sendJson(['status'=>202, 'data'=>$this->documents->getDocument($vysledok['id'])]);
+      } else {
+        $this->redirect(':Admin:Clanky:', $id);
+      }
+      
+		//} catch (Database\DriverException $e) {
 			//$button->addError($e->getMessage());
-		}
+		//}
   }
 
   
   /**
    * Upload prilohy
-   * @param ArrayHash $values
    * @return array */
-  private function _uploadPriloha(ArrayHash $values): array {
-    //dumpe($values);
-    $pr = $this->documents->find($values->id);//Zmazanie starej prílohy
-    if ($pr != null) {
-      if (is_file($pr->main_file)) { unlink($this->wwwDir."/".$pr->main_file);}
-      if (is_file($pr->thumb_file)) { unlink($this->wwwDir."/".$pr->thumb_file);}
+  private function _uploadPriloha(int $id, array $values): array {
+    if ($id) { // Hladám, len ak mením
+      $pr = $this->documents->find($id);
+      if ($pr != null) { //Zmazanie starých súborov prílohy
+        if (is_file($pr->main_file)) { unlink($this->wwwDir."/".$pr->main_file);}
+        if (is_file($pr->thumb_file)) { unlink($this->wwwDir."/".$pr->thumb_file);}
+      }
     }
-    $fileName = $values->priloha->getSanitizedName();
-		$pi = pathinfo($fileName);
-		$additionalToken = 0;
-		//Najdi meno suboru
-		if (file_exists($this->nastavenie['prilohy_dir'].$fileName)) {
-			do { $additionalToken++;
-			} while (file_exists($this->nastavenie['prilohy_dir'].$pi['filename'].$additionalToken.".".$pi['extension']));
-    }
-		$finalFileName = ($additionalToken == 0) ? $pi['filename'] : $pi['filename'].$additionalToken;
+    
+    // Vytvor bezkonfliktné meno súboru na uloženie
+    $fname = $this->_getFileName($values['priloha'], $this->nastavenie['prilohy_dir']);
+
+    //dumpe($values['priloha'], $fname);
 		//Presun subor na finalne miesto a ak je to obrazok tak vytvor nahlad
-		$values->priloha->move($this->nastavenie['prilohy_dir'].$finalFileName.".". $pi['extension']);
-		if ($values->priloha->isImage()) {  // Ak nahravam obrazok
-			$image_name = $this->nastavenie['prilohy_dir'].$finalFileName.".". $pi['extension'];
-			$thumb_name = $this->nastavenie['prilohy_dir'].'tb_'.$finalFileName.".". $pi['extension'];
+		$values['priloha']->move($this->nastavenie['prilohy_dir'].$fname['ename']);
+		if ($values['priloha']->isImage()) {  // Ak nahravam obrazok
+			$image_name = $this->nastavenie['prilohy_dir'].$fname['ename'];
+			$thumb_name = $this->nastavenie['prilohy_dir'].'tb_'.$fname['ename'];
 			$image = Image::fromFile($image_name);
       $image->resize($this->nastavenie['prilohy_images']['x'], $this->nastavenie['prilohy_images']['y'], Image::SHRINK_ONLY);
       $image->save($image_name, $this->nastavenie['prilohy_images']['kvalita']);
@@ -121,48 +131,67 @@ class DokumentyPresenter extends BasePresenter {
 			$thumb->resize($this->nastavenie['prilohy_images']['tx'], $this->nastavenie['prilohy_images']['ty'], Image::SHRINK_ONLY);// | Image::EXACT
 			$thumb->save($thumb_name, $this->nastavenie['prilohy_images']['tkvalita']);
     } else { // Ak nahravam subor
-      if ($values->thumb->hasFile() && $values->thumb->isImage()) { // Ak mam nahlad
-        $thumbInfo = pathinfo($values->thumb->name);
-        $thumb_name = $this->nastavenie['prilohy_dir'].'tb_'.$finalFileName.".".$thumbInfo['extension']; 
-        $values->thumb->move($thumb_name);
+      if ($values['thumb']->hasFile() && $values['thumb']->isImage()) { // Ak mam nahlad
+        $thumbInfo = pathinfo($values['thumb']->name);
+        $thumb_name = $this->nastavenie['prilohy_dir'].'tb_'.$fname['ename']; 
+        $values['thumb']->move($thumb_name);
         $thumb = Image::fromFile($thumb_name);
         $thumb->resize($this->nastavenie['prilohy_images']['tx'], $this->nastavenie['prilohy_images']['ty'], Image::SHRINK_ONLY);// | Image::EXACT
         $thumb->save($thumb_name, $this->nastavenie['prilohy_images']['tkvalita']);
       } else { // Ak nemam nehlad
-        $thumb_name = is_file($this->wwwDir. "/ikonky/Free-file-icons-master/48px/". $pi['extension'].".png") ? "www/ikonky/Free-file-icons-master/48px/". $pi['extension'].".png" : NULL;
+        $thumb_name = is_file($this->wwwDir. "/ikonky/Free-file-icons-master/48px/". $fname['extension'].".png") ? "www/ikonky/Free-file-icons-master/48px/". $pi['extension'].".png" : NULL;
       }
     }
   
 		return [
-			'finalFileName' => $finalFileName,
-			'pripona'				=> $pi['extension'],
+			'finalFileName' => $fname['name'],
+			'pripona'				=> $fname['extension'],
 			'thumb'					=> isset($thumb_name) ? $thumb_name : NULL,
-      'is_image'      => $values->priloha->isImage()  
+      'is_image'      => $values['priloha']->isImage()  
 		];
   }
   
   /**
    * Upload nahladu
-   * @param FileUpload $values
-   * @return array */
-  private function _uploadThumb(FileUpload $values): array {
-    $pr = $this->dokumenty->find($values->id);//Zmazanie starej prílohy
-    if ($pr !== FALSE) {
-      if (is_file($pr->thumb_file)) { unlink($this->wwwDir."/".$pr->thumb_file);}
+   * @param int $id Id dokumentu. Akje 0 tak sa pridáva
+   * @param FileUpload $thumb Súbor náhľadu.
+   * @return String Kompletný názov uloženého súboru aj s reltívnou cestou */
+  private function _uploadThumb(int $id, FileUpload $thumb): String {
+    if ($id) { // Mazanie sa vykoná len ak je $id > 0
+      $pr = $this->dokumenty->find($id);//Zmazanie starej prílohy
+      if ($pr !== FALSE) {
+        if (is_file($pr->thumb_file)) { unlink($this->wwwDir."/".$pr->thumb_file);}
+      }
     }
-    $main_file = pathinfo($pr->main_file);
+    
+    // Vytvor bezkonfliktné meno súboru na uloženie
+    $fname = $this->_getFileName($thumb, $this->nastavenie['prilohy_dir'], 'tb_');
+
+    // Uloženie náhľadu podľa paametrov
+    $thumb->move($this->nastavenie['prilohy_dir'].$fname['ename']);
+    $thumb = Image::fromFile($this->nastavenie['prilohy_dir'].$fname['ename']);
+    $thumb->resize($this->nastavenie['prilohy_images']['tx'], $this->nastavenie['prilohy_images']['ty'], Image::SHRINK_ONLY);// | Image::EXACT
+    $thumb->save($this->nastavenie['prilohy_dir'].$fname['ename'], $this->nastavenie['prilohy_images']['tkvalita']);
+  
+		return $this->nastavenie['prilohy_dir'].$fname['ename'];
+  }
+
+  /** 
+   * Vytvorenie bezpečného mena súboru 
+   * @return array ['name'=>'meno súboru', 'ename'=>'meno súboru s príponou', 'extension'=>'prípona'] */
+  private function _getFileName(FileUpload $file, String $dir, String $prefix = ""): array {
+    $file_info = pathinfo($file->getSanitizedName());
     $additionalToken = 0;
 		//Najdi meno suboru
-		if (file_exists($this->nastavenie['prilohy_dir'].'tb_'.$main_file['filename'].'.jpg')) {
+		if (file_exists($dir.$prefix.$file_info['filename'].$file_info['extension'])) {
 			do { $additionalToken++;
-			} while (file_exists($this->nastavenie['prilohy_dir'].'tb_'.$main_file['filename'].$additionalToken.".jpg"));
+			} while (file_exists($dir.$prefix.$file_info['filename'].$additionalToken.$file_info['extension']));
     }
-		$finalFileName = 'tb_'.$main_file['filename'].($additionalToken == 0 ? '' : $additionalToken).".jpg";
-    $values->thumb->move($this->nastavenie['prilohy_dir'].$finalFileName);
-    $thumb = Image::fromFile($this->nastavenie['prilohy_dir'].$finalFileName);
-    $thumb->resize($this->nastavenie['prilohy_images']['tx'], $this->nastavenie['prilohy_images']['ty'], Image::SHRINK_ONLY);// | Image::EXACT
-    $thumb->save($this->nastavenie['prilohy_dir'].$finalFileName, $this->nastavenie['prilohy_images']['tkvalita']);
-  
-		return $this->nastavenie['prilohy_dir'].$finalFileName;
+	
+    return [
+      'name'=> $prefix.$file_info['filename'].($additionalToken == 0 ? '' : $additionalToken),
+      'ename' => $prefix.$file_info['filename'].($additionalToken == 0 ? '' : $additionalToken).".".$file_info['extension'],
+      'extension' => $file_info['extension']
+    ];
   }
 }
